fix(LAS-131): address greptile-apps review concerns
- Extract SubtaskBadge, SubtaskList, SubtaskCounts, buildSubtaskCountMap
  to subtask-utils.tsx — breaks circular import between IssuesList and KanbanBoard
- Fix O(n²) scan in highlightedSubtaskIds: build issueById Map once (O(n))
  instead of calling issues.find() per filtered ID
- Fix buildRootAndSubtaskMap mismatch: pass sourceIssues (same source used
  to build filtered) as allIssues so search results with uncached parents
  are not silently dropped
